Python/requirements.txt
# requirements.txt

# Numerical computing library
numpy==1.21.2

# Python Imaging Library (PIL) fork
Pillow==8.3.2
